requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
sqlalchemy-migrate==0.11.0